Moved repositories and viewmodels into the db package.
